--- Content.Server/Tabletop/TabletopHologramComponent.cs
@@ -1,10 +1,0 @@
-namespace Content.Server.Tabletop;
-
-/// <summary>
-/// This is used for tracking pieces that are simply "holograms" shown on the tabletop
-/// </summary>
-[RegisterComponent]
-public sealed partial class TabletopHologramComponent : Component
-{
-
-}
