--- Content.Server/Tabletop/TabletopSessionPlayerData.cs
@@ -1,10 +1,0 @@
-namespace Content.Server.Tabletop
-{
-    /// <summary>
-    ///     A class that stores per-player data for tabletops.
-    /// </summary>
-    public sealed class TabletopSessionPlayerData
-    {
-        public EntityUid Camera { get; set; }
-    }
-}
