+++ Content.Shared/Tabletop/TabletopSessionPlayerData.cs
@@ -1,0 +1,9 @@
+namespace Content.Shared.Tabletop;
+
+/// <summary>
+/// A class that stores per-player data for tabletops.
+/// </summary>
+public sealed class TabletopSessionPlayerData
+{
+    public EntityUid Camera { get; set; }
+}
